refactor(srcs): update architecture using classes for server
