+++ srcs/check.cpp
@@ -1,0 +1,61 @@
+#include "check.hpp"
+
+bool only_digits(char *str)
+{
+    for (unsigned int i = 0; i < strlen(str); i++)
+    {
+        if (!isdigit(str[i]))
+        {
+            return (false);
+        }
+    }
+    return (true);
+}
+
+int check_args(int argc, char **argv)
+{
+    if (argc != 3)
+    {
+        std::cout << "Error: invalid number of arguments\n";
+        return (-1);
+    }
+    else
+    {
+        if (!only_digits(argv[1]))
+        {
+            std::cout << "Error: invalid port (only digits)\n";
+            return (-1);
+        }
+        else if (atoi(argv[1]) < 1024)
+        {
+            std::cout << "Error: invalid port (1024 minimum)\n";
+            return (-1);
+        }
+    }
+    return (0);
+}
+
+void    check_password(users *all_users, pollfd *fds, std::string *value, char *password, unsigned int i)
+{
+    if (get_user(i - 1, all_users)->connected)
+    {
+        send(fds[i].fd, "ERR_ALREADYREGISTRED\n", strlen("ERR_ALREADYREGISTRED\n"), 0);
+    }
+    else if (value[1].empty())
+    {
+        send(fds[i].fd, "ERR_NEEDMOREPARAMS\n", strlen("ERR_NEEDMOREPARAMS\n"), 0);
+    }
+    else if (!value[2].empty())
+    {
+        send(fds[i].fd, "Error: invalid password (no spaces)\n", strlen("Error: invalid password (no spaces)\n"), 0);
+    }
+    else if (value[1] != password)
+    {
+        send_err_password(fds[i].fd);
+    }
+    else
+    {
+        get_user(i - 1, all_users)->connected = true;
+        send(fds[i].fd, "Password OK\n", strlen("Password OK\n"), 0);
+    }
+}
