+++ srcs/response.cpp
@@ -18,3 +18,15 @@ void	send_err_password(int socket)
 	std::string message = ":127.0.0.1 464 ERR_PASSWDMISMATCH :Password incorrect\r\n";
 	send(socket, message.c_str(), message.length(), 0);
 }
+
+void    send_broadcast(unsigned int nb_fd, struct pollfd *fds, std::string message)
+{
+    unsigned int i = 1;
+
+    while (i < nb_fd)
+    {
+        std::cout << "sending to: " << i << std::endl;
+        send(fds[i].fd, message.c_str(), strlen(message.c_str()), 0);
+        i++;
+    }
+}
